requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pyTelegramBotAPI==4.27.0
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
python-dotenv==1.1.1